Move async/await code into Fluent.swift
