+++ Sources/HummingbirdFluent/Fluent.swift
@@ -112,3 +112,17 @@ public struct HBFluent {
             )!
     }
 }
+
+/// async/await
+@available(macOS 12.0, iOS 15.0, watchOS 8.0, tvOS 15.0, *)
+extension HBFluent {
+    /// Run migration if needed
+    public func migrate() async throws {
+        try await self.migrate().get()
+    }
+
+    /// Run revert if needed
+    public func revert() async throws {
+        try await self.revert().get()
+    }
+}
